pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0", "wheel>=0.42.0"]
build-backend = "setuptools.build_meta"

[project]
name = "travel_planner"
version = "0.1.0"
description = "Multi-agent travel planning system powered by OpenAI Agents SDK and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "Bjorn Melin", email = "bjorn@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai-agents>=0.0.14", # OpenAI Agents SDK (latest version as of May 2025)
    "langchain>=0.1.0",
    "langgraph>=0.4.0",
    "stagehand-py>=0.3.6", # Browser automation (latest version as of May 2025)
    "supabase>=2.1.0",
    "playwright>=1.42.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.9.0",
    "loguru>=0.7.2",
    "tenacity>=8.2.3",
    "httpx>=0.26.0",
    "python-dateutil>=2.8.2",
    "pycountry>=23.12.11",
    "tavily-python>=0.2.8",
    "firecrawl>=0.1.0",
    "agents-sdk-models>=0.0.11", # Model adapters for OpenAI Agents SDK
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "freezegun>=1.4.0",
    "respx>=0.20.2",
    "black>=24.1.0",
    "isort>=5.13.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-python-dateutil>=2.8.19.14",
    "types-requests>=2.31.0.2",
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=1.3.0",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
